fix: move video batch-load from observer to ListProduct plugin
  The former catalog_block_product_list_collection observer called
  $collection->getSize() and foreach, forcing a collection load before
  the toolbar applied pageSize/curPage. Under Amasty Shopby AJAX
  pagination this returned the full category (all products) instead of
  the requested page.

  Replace the observer with a plugin on ListProduct::getLoadedProductCollection
  that guards on $collection->isLoaded() so the batch video enrichment
  only runs after the toolbar-applied pagination is in effect. Only the
  products on the current page are processed.

  Fixes category AJAX pagination on stores using Amasty Shopby + videos
  enabled on listings (rollpix_gallery/video/listing_enabled=1).
