--- etc/frontend/events.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Rollpix ProductGallery Frontend Events
- *
- * @category  Rollpix
- * @package   Rollpix_ProductGallery
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:noNamespaceSchemaLocation="urn:magento:framework:Event/etc/events.xsd">
-
-    <!-- Pre-load video data for product listing collections -->
-    <event name="catalog_block_product_list_collection">
-        <observer name="rollpix_add_video_data_to_collection"
-                  instance="Rollpix\ProductGallery\Observer\AddVideoDataToCollection"/>
-    </event>
-
-</config>
